repaired GLubyte bug
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@32 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/MapEditorModule.h
@@ -29,7 +29,7 @@ class MapEditorModule : public Module
     private:
         Matrix<int> mTerrainHeight;
         GLfloat* mTerrainVertices;
-        GLubyte* mTerrainIndices;
+        GLuint* mTerrainIndices;
 
         TSphere mSphere;
         GLsizei mNumIndices;
